requirements.txt
Django>=5.0,<6.0
python-decouple>=3.8
Pillow>=10.0.0
psycopg2-binary>=2.9.10,<3.0.0
djangorestframework>=3.15.0
django-cors-headers>=4.3.0
djangorestframework-simplejwt>=5.3.0
whitenoise>=6.6.0
gunicorn>=21.2.0
dj-database-url>=2.1.0
